增加了跳转当前音乐项，优化了音量控制交互体验。 修复了部分Bug，添加了更多Bug。 回放增益默认开启，会导致音频初次加载缓慢~

+++ QwQ Music/Models/ConfigInfoModel.cs
@@ -9,5 +9,5 @@ public class ConfigInfoModel
 
 public class PlayerConfig
 {
-    public float Volume { get; init; }
+    public int Volume { get; init; }
 }
